fixed bug causing the script to not work when ran not from the project directory; added new functionalities

+++ merger.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 import sys
+import os
 from pathlib import Path
 
 
@@ -16,7 +17,7 @@ class Dependency:
         return f"{self.path}"
 
 
-def check_file(path, dependencies, cwd = Path.cwd()):
+def check_file(path, dependencies, cwd):
     new_dependency_found = False
     with open(path) as f:
         include_lines = [line.split()[1].replace('"','') for line in f.read().splitlines() if line.startswith("#include") and '"' in line]
@@ -25,21 +26,21 @@ def check_file(path, dependencies, cwd = Path.cwd()):
         dependency = Dependency((cwd / Path(line)).resolve(strict=False))
         if dependency.name not in [d.name for d in dependencies]:
             dependencies.append(dependency)
-            new_dependency_found = True
-    if new_dependency_found:
-        for dep in dependencies:
-            check_file(dep.path, dependencies, dep.path.parent)
+            check_file(dependency.path, dependencies, dependency.path.parent)
 
 
 def main():
     if len(sys.argv) < 3:
-        raise Exception(f"usage: {sys.argv[0]} src_file out_file")
-    src_path = sys.argv[1]
-    out_path = sys.argv[2]
+        raise Exception(f"usage: {sys.argv[0]} src_file ... out_file")
+    src_path = sys.argv[1:-1]
+    out_path = sys.argv[-1]
 
-    dependencies = [Dependency((Path.cwd() / Path(src_path)).resolve(strict=False))]
-
-    check_file(src_path, dependencies)
+    dependencies = []
+    for path in src_path:
+        dependency = Dependency((Path.cwd() / Path(path)).resolve(strict=False))
+        if dependency.name not in [d.name for d in dependencies]:
+            dependencies.append(Dependency((Path.cwd() / Path(path)).resolve(strict=False)))
+            check_file(path, dependencies, dependency.path.parent)
 
     # prepare the out file content
     out_file_content = ""
@@ -54,8 +55,13 @@ def main():
         out_file_content += "\n"
 
     # write the out file content to the file
-    with open(out_path, "w", encoding="utf8") as f:
-        f.write(out_file_content)
+    choice = "y"
+    if os.path.isfile(out_path):
+        print(f"{out_path} exists and will be overwritten!")
+        choice = input("Do you want to continue?[y/n]> ")
+    if choice.lower() == "y":
+        with open(out_path, "w", encoding="utf8") as f:
+            f.write(out_file_content)
 
 
 if __name__ == "__main__":
